data-server/requirements.txt
# Flask and Web Framework
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Werkzeug==2.3.6

# Database
psycopg2-binary==2.9.7
alembic==1.11.1

# HTTP Clients and Parsing
requests==2.31.0
urllib3==2.0.4
beautifulsoup4==4.12.2

# Data Processing and Analysis
pandas==2.0.3
numpy==1.25.1
scikit-learn==1.3.0

# Configuration and Environment
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging and Monitoring
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Security and Validation
marshmallow==3.20.1
bcrypt==4.0.1

# Date and Time
python-dateutil==2.8.2

# Regular Expressions and Text Processing
regex==2023.6.3

# Command Line Interface
click==8.1.6